Extract world from steps
